# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API Tierra en Calma
  version: 1.0.0
  description: Endpoints de usuarios y plantas

servers:
  - url: http://localhost:3001/api

paths:
  /register:
    post:
      summary: Registrar usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_usuario:
                  type: integer
                nombre:
                  type: string
                apellido:
                  type: string
                telefono:
                  type: string
                correo_electronico:
                  type: string
                contrasena:
                  type: string
              required:
                - id_usuario
                - nombre
                - correo_electronico
                - contrasena
      responses:
        "200":
          description: Usuario registrado con éxito

  /login:
    post:
      summary: Login usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                correo_electronico:
                  type: string
                contrasena:
                  type: string
              required:
                - correo_electronico
                - contrasena
      responses:
        "200":
          description: Login exitoso
        "401":
          description: Credenciales inválidas

  /registrar-planta:
    post:
      summary: Registrar planta en jardín de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_usuario:
                  type: integer
                id_planta:
                  type: integer
              required:
                - id_usuario
                - id_planta
      responses:
        "200":
          description: Planta añadida con éxito

  /health:
    get:
      tags:
        - Health
      summary: Healthcheck del servicio
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
              examples:
                ok:
                  value:
                    ok: true

  /health/db:
    get:
      tags:
        - Health
      summary: Ping a la base de datos
      responses:
        "200":
          description: Conexión a DB OK
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                properties:
                  ok:
                    type: boolean
              examples:
                ok:
                  value:
                    ok: true
        "500":
          description: Error conectando a DB
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - error
                properties:
                  ok:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "timeout"
              examples:
                fail:
                  value:
                    ok: false
                    error: "timeout"